reorganizing some commons methods
